+++ test/change_health/response/response_data_test.rb
@@ -1,7 +1,8 @@
+# frozen_string_literal: true
+
 require 'test_helper'
 
 class ResponseDataTest < Minitest::Test
-
   class FakeData < ChangeHealth::Response::ResponseData
   end
 
@@ -12,6 +13,10 @@ class ResponseDataTest < Minitest::Test
 
     def parsed_response
       @response.call
+    end
+
+    def code
+      '500'
     end
   end
 
@@ -24,7 +29,7 @@ class ResponseDataTest < Minitest::Test
       end
 
       it 'can take a response' do
-        fake_data = FakeData.new(response: FakeResponse.new(->() { 'hi' }))
+        fake_data = FakeData.new(response: FakeResponse.new(-> { 'hi' }))
         assert(false == fake_data.nil?)
       end
 
@@ -34,17 +39,16 @@ class ResponseDataTest < Minitest::Test
       end
 
       it 'handles bad response gracefully' do
-        fake_data = FakeData.new(response: FakeResponse.new(->() { JSON.parse('bad json') }))
+        fake_data = FakeData.new(response: FakeResponse.new(-> { JSON.parse('bad json') }))
         assert(false == fake_data.nil?)
       end
 
       it 'does not handle other errors gracefully' do
-        assert_raises { FakeData.new(response: FakeResponse.new(->() { nil.say_hi! })) }
+        assert_raises { FakeData.new(response: FakeResponse.new(-> { nil.say_hi! })) }
       end
     end
 
     describe 'error handling' do
-
       describe '#errors?' do
         it 'false if none' do
           assert_equal(false, fake_data.errors?)
@@ -54,21 +58,47 @@ class ResponseDataTest < Minitest::Test
           assert_equal(false, fake_data.recommend_retry?)
         end
 
-        describe 'with errors' do
-          let(:json_data) { load_sample('error_response.fields.json', parse: true) }
+        describe 'error' do
+          let(:fake_data) do
+            FakeData.new(
+              data: json_data,
+              response: FakeResponse.new(-> { json_data })
+            )
+          end
+          describe 'with errors' do
+            let(:json_data) { load_sample('error_response.fields.json', parse: true) }
 
-          it 'true if errors' do
-            assert_equal(true, fake_data.errors?)
+            it 'true if errors' do
+              assert_equal(true, fake_data.errors?)
+              assert_equal(5, fake_data.errors.size)
+              assert_nil(fake_data.server_error)
+            end
+          end
+
+          describe 'server_error' do
+            let(:json_data) { load_sample('error_response.json', parse: true) }
+
+            it 'true if server_error' do
+              assert_equal(true, fake_data.errors?)
+              assert_equal(0, fake_data.errors.size)
+              assert_equal('Failed Request: HTTP code: 500 MSG: Invalid request', fake_data.server_error.message)
+            end
           end
         end
       end
 
       describe 'more complicated errors' do
-        let(:field_error0) { {'field' => 'patient.name', 'description' => 'is too short' } }
-        let(:field_error1) { {'field' => 'cat', 'description' => 'has meow' } }
-        let(:field_down) { {'field' => 'Http Header', 'description' => 'Please review http headers for this API, please contact support if you are unsure how to resolve.'} }
-        let(:code_needs_fix) { {'code' => '71', 'description' => 'Need more time' } }
-        let(:code_retry_80) { {'code' => '80', 'description' => 'Unable to Respond at Current Time', 'followupAction' => 'Resubmission Allowed'} }
+        let(:field_error0) { { 'field' => 'patient.name', 'description' => 'is too short' } }
+        let(:field_error1) { { 'field' => 'cat', 'description' => 'has meow' } }
+        let(:field_down) do
+          { 'field' => 'Http Header',
+            'description' => 'Please review http headers for this API, please contact support if you are unsure how to resolve.' }
+        end
+        let(:code_needs_fix) { { 'code' => '71', 'description' => 'Need more time' } }
+        let(:code_retry_80) do
+          { 'code' => '80', 'description' => 'Unable to Respond at Current Time',
+            'followupAction' => 'Resubmission Allowed' }
+        end
         let(:code_noretry_80) { code_retry_80.merge('followupAction' => 'xxDo Not Resubmitmm;') }
         let(:json_data) { { 'errors' => errors } }
 
@@ -97,8 +127,8 @@ class ResponseDataTest < Minitest::Test
           end
 
           it 'has messages from fields' do
-            assert_equal("patient.name: is too short", fake_data.errors[0].message)
-            assert_equal("cat: has meow", fake_data.errors[1].message)
+            assert_equal('patient.name: is too short', fake_data.errors[0].message)
+            assert_equal('cat: has meow', fake_data.errors[1].message)
           end
         end
 
@@ -173,4 +203,3 @@ class ResponseDataTest < Minitest::Test
     end
   end
 end
-